# libraries/archive/react-components/api/internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Idol Promo API
  version: 1.0.0
  description: Go Fiber API for authentication, items, products, and uploads.
servers:
  - url: /
    description: Same-origin frontend proxy
tags:
  - name: Health
  - name: Authentication
  - name: Items
  - name: Products
paths:
  /livez:
    get:
      tags: [Health]
      summary: Process liveness
      responses:
        "200":
          description: Process is alive
  /readyz:
    get:
      tags: [Health]
      summary: Dependency readiness
      responses:
        "200":
          description: API is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          $ref: "#/components/responses/Error"
  /startupz:
    get:
      tags: [Health]
      summary: Startup completion
      responses:
        "200":
          description: Startup completed
        "503":
          $ref: "#/components/responses/Error"
  /api/items:
    get:
      tags: [Items]
      summary: List items
      responses:
        "200":
          description: Item collection
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "500":
          $ref: "#/components/responses/Error"
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Access token and user; refresh token is set as an httpOnly cookie
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate the refresh token and issue a new session
      security:
        - refreshCookie: []
      responses:
        "200":
          description: Rotated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current session
      parameters:
        - in: query
          name: all
          schema: { type: boolean, default: false }
          description: Revoke every refresh token for the user
      security:
        - refreshCookie: []
      responses:
        "204": { description: Session revoked }
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Current database-backed identity
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202": { description: Accepted regardless of account existence }
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Consume a reset token and set a password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, format: password, minLength: 12, maxLength: 72 }
      responses:
        "200": { description: Password changed and sessions revoked }
        "400":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/products/:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - { in: query, name: mode, schema: { type: string, enum: [offset, keyset, all] } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 200, default: 25 } }
        - { in: query, name: cursor, schema: { type: string } }
        - { in: query, name: withTotal, schema: { type: boolean, default: true } }
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: sortBy, schema: { type: string } }
        - { in: query, name: sortDir, schema: { type: string, enum: [asc, desc] } }
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: condition, schema: { type: string } }
        - { in: query, name: availability, schema: { type: string } }
        - { in: query, name: isPublished, schema: { type: boolean } }
        - { in: query, name: tag, schema: { type: string } }
        - { in: query, name: region, schema: { type: string } }
      responses:
        "200":
          description: Product page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Products]
      summary: Create a product
      description: Requires the products:create capability.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          schema: { type: string, minLength: 8, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/ProductInput"
                - type: object
                  properties:
                    productImage: { type: string, format: binary }
                    productDocuments:
                      type: array
                      items: { type: string, format: binary }
      responses:
        "201":
          description: Product created
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/products/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    get:
      tags: [Products]
      summary: Get a product
      responses:
        "200":
          description: Product
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Products]
      summary: Partially update a product
      description: Requires the products:update capability.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: If-Match
          schema: { type: string }
          description: Optional ETag/version returned by a previous read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/ProductPatch"
                - type: object
                  properties:
                    productImage: { type: string, format: binary }
                    productDocuments:
                      type: array
                      items: { type: string, format: binary }
      responses:
        "200":
          description: Updated product
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "412":
          $ref: "#/components/responses/Error"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Products]
      summary: Delete a product
      description: Requires the products:delete capability.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: If-Match
          schema: { type: string }
      responses:
        "204": { description: Product deleted }
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "412":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: idol_refresh
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, example: ok }
    Item:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
    User:
      type: object
      required: [id, email, displayName, role]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [admin, user] }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        rememberMe: { type: boolean, default: false }
    Session:
      type: object
      required: [token, expiresIn, user]
      properties:
        token: { type: string }
        expiresIn: { type: integer, example: 900 }
        user: { $ref: "#/components/schemas/User" }
    ProductFields:
      type: object
      properties:
        productId: { type: string, minLength: 3, maxLength: 15 }
        productName: { type: string, minLength: 2, maxLength: 100 }
        description: { type: string, maxLength: 500 }
        price: { type: string, pattern: "^[0-9]+(\\.[0-9]{1,2})?$" }
        rating: { type: integer, minimum: 0, maximum: 5 }
        comments: { type: [string, "null"] }
        releaseDate: { type: string, format: date }
        supportEmail: { type: [string, "null"], format: email }
        supportPhone: { type: [string, "null"] }
        productUrl: { type: [string, "null"], format: uri }
        themeColor: { type: [string, "null"] }
        condition: { type: string, enum: [new, refurbished, used] }
        availability: { type: string, enum: [inStock, preorder, discontinued], default: inStock }
        tags: { type: array, items: { type: string } }
        shippingRegions: { type: array, minItems: 1, items: { type: string } }
        warrantyMonths: { type: integer, minimum: 0, maximum: 120 }
        isPublished: { type: boolean, default: false }
        acceptTerms: { type: boolean, const: true }
        category: { type: string, enum: [electronics, clothing, books, home] }
    ProductInput:
      allOf:
        - $ref: "#/components/schemas/ProductFields"
        - type: object
          required: [productId, productName, description, price, releaseDate, condition, shippingRegions, acceptTerms, category]
    ProductPatch:
      $ref: "#/components/schemas/ProductFields"
    Product:
      allOf:
        - $ref: "#/components/schemas/ProductFields"
        - type: object
          required: [id, version, createdAt, updatedAt]
          properties:
            id: { type: integer, format: int64 }
            version: { type: integer, format: int64, minimum: 1 }
            productImagePath: { type: [string, "null"] }
            productDocumentPaths: { type: array, items: { type: string } }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    ProductPage:
      type: object
      required: [items, total, page, pageSize, hasMore]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Product" } }
        total: { type: [integer, "null"], format: int64 }
        page: { type: integer }
        pageSize: { type: integer }
        hasMore: { type: boolean }
        nextCursor: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }